# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Altera Echoes of Fate API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OkResponse {
  ok: boolean;
}

export type RaceBonuses = { [key: string]: number };

export interface Race {
  key: string;
  nameRu: string;
  nameEn: string;
  lore: string;
  creatureForm?: string | null;
  bonuses: RaceBonuses;
}

export type CharClassInfoStartStats = { [key: string]: number };

export interface CharClassInfo {
  key: string;
  nameRu: string;
  desc: string;
  resource: string;
  startStats: CharClassInfoStartStats;
}

export interface Location {
  id: string;
  name: string;
  region: string;
  description: string;
  type: string;
  isSafe: boolean;
}

export interface EnemyTemplate {
  key: string;
  name: string;
  level: number;
  hp: number;
  damage: number;
  rewardSilver: number;
  rewardExp: number;
  locationId: string;
  lore: string;
}

export interface LoreResponse {
  races: Race[];
  classes: CharClassInfo[];
  locations: Location[];
  enemies: EnemyTemplate[];
}

export interface Character {
  id: number;
  name: string;
  race: string;
  charClass: string;
  level: number;
  experience: number;
  nextLevelExp: number;
  hp: number;
  maxHp: number;
  mana: number;
  maxMana: number;
  energy: number;
  maxEnergy: number;
  strength: number;
  agility: number;
  intelligence: number;
  endurance: number;
  intuition: number;
  luck: number;
  statPoints: number;
  influencePoints: number;
  silver: number;
  locationId: string;
  inBattle: boolean;
  isAlive: boolean;
  totalKills: number;
  totalDeaths: number;
}

export interface CharacterResponse {
  character: Character | null;
}

export interface CreateCharacterRequest {
  /**
   * @minLength 2
   * @maxLength 30
   */
  name: string;
  race: string;
  charClass: string;
}

export type AllocateStatRequestStat =
  (typeof AllocateStatRequestStat)[keyof typeof AllocateStatRequestStat];

export const AllocateStatRequestStat = {
  strength: "strength",
  agility: "agility",
  intelligence: "intelligence",
  endurance: "endurance",
  intuition: "intuition",
} as const;

export interface AllocateStatRequest {
  stat: AllocateStatRequestStat;
}

export interface MoveRequest {
  locationId: string;
}

export type BattleStatus = (typeof BattleStatus)[keyof typeof BattleStatus];

export const BattleStatus = {
  active: "active",
  victory: "victory",
  defeat: "defeat",
  fled: "fled",
} as const;

export type BattleLogEntryActor =
  (typeof BattleLogEntryActor)[keyof typeof BattleLogEntryActor];

export const BattleLogEntryActor = {
  player: "player",
  enemy: "enemy",
  system: "system",
} as const;

export interface BattleLogEntry {
  round: number;
  actor: BattleLogEntryActor;
  action: string;
  text: string;
  damage?: number;
  crit?: boolean;
}

export interface Battle {
  id: number;
  characterId: number;
  enemyKey: string;
  enemyName: string;
  enemyLevel: number;
  enemyHp: number;
  enemyMaxHp: number;
  characterHp: number;
  status: BattleStatus;
  log: BattleLogEntry[];
  rewardSilver: number;
  rewardExp: number;
}

export interface BattleResponse {
  battle: Battle | null;
}

export interface StartBattleRequest {
  enemyKey: string;
}

export type BattleActionRequestAction =
  (typeof BattleActionRequestAction)[keyof typeof BattleActionRequestAction];

export const BattleActionRequestAction = {
  attack: "attack",
  heavy: "heavy",
  defend: "defend",
  flee: "flee",
} as const;

export interface BattleActionRequest {
  action: BattleActionRequestAction;
}

export interface NPCWithReputation {
  id: string;
  name: string;
  title?: string | null;
  role: string;
  locationId: string;
  shortProfile: string;
  voiceStyle: string;
  faction: string;
  reputation: number;
  repName: string;
  repIcon: string;
  repBehavior: string;
  tier: number;
}

export type DialogueMessageRole =
  (typeof DialogueMessageRole)[keyof typeof DialogueMessageRole];

export const DialogueMessageRole = {
  player: "player",
  npc: "npc",
  system: "system",
} as const;

export interface DialogueMessage {
  id: number;
  role: DialogueMessageRole;
  content: string;
  createdAt: string;
}

export type SendDialogueRequestTone =
  (typeof SendDialogueRequestTone)[keyof typeof SendDialogueRequestTone];

export const SendDialogueRequestTone = {
  neutral: "neutral",
  friendly: "friendly",
  flirt: "flirt",
  threat: "threat",
  insult: "insult",
  persuade: "persuade",
} as const;

export interface SendDialogueRequest {
  /**
   * @minLength 1
   * @maxLength 500
   */
  content: string;
  tone: SendDialogueRequestTone;
}

export interface DialogueResponse {
  npcReply: string;
  repDelta: number;
  newReputation: number;
  repName: string;
  eventTriggered: string | null;
}

export interface GiftRequest {
  /** @minimum 1 */
  silver: number;
}

export interface GiftResponse {
  reputation: number;
  repName: string;
  silverLeft: number;
  message: string;
}

export interface LedgerEntry {
  id: number;
  eventType: string;
  targetId?: string | null;
  locationId?: string | null;
  description: string;
  deltaRep: number;
  severity: number;
  isPublic: boolean;
  createdAt: string;
}

export type InventoryItemStats = { [key: string]: unknown };

export interface InventoryItem {
  id: number;
  itemKey: string;
  name: string;
  itemType: string;
  rarity: string;
  quantity: number;
  equipped: boolean;
  stats: InventoryItemStats;
}

export interface EquipRequest {
  itemId: number;
}

export type ListNpcsParams = {
  locationId?: string;
};