# lib/api-zod/src/generated/types/battleActionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Altera Echoes of Fate API
 * OpenAPI spec version: 0.1.0
 */
import type { BattleActionRequestAction } from "./battleActionRequestAction";

export interface BattleActionRequest {
  action: BattleActionRequestAction;
}