# lib/api-zod/src/generated/types/character.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Altera Echoes of Fate API
 * OpenAPI spec version: 0.1.0
 */

export interface Character {
  id: number;
  name: string;
  race: string;
  charClass: string;
  level: number;
  experience: number;
  nextLevelExp: number;
  hp: number;
  maxHp: number;
  mana: number;
  maxMana: number;
  energy: number;
  maxEnergy: number;
  strength: number;
  agility: number;
  intelligence: number;
  endurance: number;
  intuition: number;
  luck: number;
  statPoints: number;
  influencePoints: number;
  silver: number;
  locationId: string;
  inBattle: boolean;
  isAlive: boolean;
  totalKills: number;
  totalDeaths: number;
}