# lib/api-zod/src/generated/types/dialogueMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Altera Echoes of Fate API
 * OpenAPI spec version: 0.1.0
 */
import type { DialogueMessageRole } from "./dialogueMessageRole";

export interface DialogueMessage {
  id: number;
  role: DialogueMessageRole;
  content: string;
  createdAt: Date;
}