# lib/api-zod/src/generated/types/giftRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Altera Echoes of Fate API
 * OpenAPI spec version: 0.1.0
 */

export interface GiftRequest {
  /** @minimum 1 */
  silver: number;
}