# lib/api-zod/src/generated/types/ledgerEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Altera Echoes of Fate API
 * OpenAPI spec version: 0.1.0
 */

export interface LedgerEntry {
  id: number;
  eventType: string;
  targetId?: string | null;
  locationId?: string | null;
  description: string;
  deltaRep: number;
  severity: number;
  isPublic: boolean;
  createdAt: Date;
}