# lib/api-zod/src/generated/types/race.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Altera Echoes of Fate API
 * OpenAPI spec version: 0.1.0
 */
import type { RaceBonuses } from "./raceBonuses";

export interface Race {
  key: string;
  nameRu: string;
  nameEn: string;
  lore: string;
  creatureForm?: string | null;
  bonuses: RaceBonuses;
}